Python BackEnd/requirements.txt
fastapi
uvicorn
prisma
google-generativeai
requests
python-dotenv
transformers
torch
sentence-transformers
ddgs

